requirements.txt
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
SQLAlchemy==1.3.8
marshmallow-sqlalchemy==0.19.0
flask-marshmallow==0.10.1
Flask-HTTPAuth==3.3.0
jwt==0.6.1
pyjwt==1.7.1
psycopg2-binary==2.8.3
validate_email==1.3
SQLAlchemy-Utils==0.34.2
click==6.7
eventlet==0.23.0
Flask==1.0.2
Flask-SocketIO==3.0.1
Flask-WTF==0.14.2
greenlet==0.4.13
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-engineio==2.1.1
python-socketio==1.9.0
six==1.11.0
Werkzeug==0.14.1
WTForms==2.2.1
gunicorn==19.9.0